==> 2_data_cleaing.ipynb <==
### Data Cleaning

import pandas as pd
import numpy as np

pd.set_option("display.max_columns", None)

# Games data

Again, this data is only used for EDA

games = pd.read_csv("data/games_raw.csv", index_col="id")

games.head()

# time column is empty
games.drop("time", axis=1, inplace=True)

# Insufficient data to use 2021 games
# Also some of these games were 'in progress' when pulled from the api, so stats are incomplete
games_2021 = games[games["season"].eq(2021)].index
games.drop(games_2021, axis=0, inplace=True)

# dropping a random unfinished game
games.drop(games[games["status"].eq("8:30 PM ET")].index, axis=0, inplace=True)

# dropping unplayed games
games.drop(games[games["home_team_score"].eq(0)].index, axis=0, inplace=True)

# dropping unecessary name info
games.drop(["home_team.city", "visitor_team.city", "home_team.name", "visitor_team.name"], axis=1, inplace=True)

games.head()

games.to_csv("data/games_clean.csv")

players = pd.read_csv("data/players_raw.csv", index_col="id")

# drop superfluous team info
players.drop(["team.name", "team.division", "team.conference", "team.city", "team.abbreviation"], axis=1, inplace=True)

# reorganize columns
players = players[['first_name', 'last_name', 'position', 'team.id', 'team.full_name',
                   'height_feet', 'height_inches', 'weight_pounds']]

players

players.isnull().sum()

# Stats data

This data is used to build the model

stats = pd.read_csv("data/stats_raw.csv", index_col="id")

stats.head(2)

# drop uneeded columns
stats.drop(["game.period", "game.postseason", "game.status", "game.time", "player.height_feet", "player.height_inches",
            "player.weight_pounds", "team.abbreviation", "team.city", "team.conference", "team.division", "team.name",
            "player", "player.first_name", "player.last_name", "player.position", "team.full_name", "player.team_id"],
          axis=1, inplace=True)

# a null value generally indicates that the player did not play in that game
stats.dropna(axis=0, how="any", inplace=True)

# clean time column to get a consitent format. ("mm:ss" or "m:ss")
stats["min"] = stats["min"].astype(str)

# drop the row if the player didn't play in the game
played_0min = stats[stats["min"].eq("0:00") | stats["min"].str.startswith("0") | stats["min"].eq("")].index
stats.drop(played_0min, axis=0, inplace=True)

# Convert times like "27.0" to "27:0"
stats["min"] = stats["min"].str.replace(".",":")

# convert times like "27" to "27:00"
minutes_only_times = stats["min"][~stats["min"].str.contains(":")].index
stats["min"].loc[minutes_only_times] += ":00"


minutes = [time[0] for time in stats["min"].str.split(":").values]
seconds = [time[1] for time in stats["min"].str.split(":").values]

# convert times like "27:0" to "27:00"
for i, second in enumerate(seconds):
    if len(second) == 1:
        seconds[i] = second + "0"

# convert times like "8:60" to "9:00"
for i, second in enumerate(seconds):        
    if second == "60":
        seconds[i] = "00"
        minutes[i] = str(int(minutes[i]) + 1)  # increment minutes by 1

times = [":".join(list(item)) for item in list(zip(minutes,seconds))]

stats["min"] = times

stats

stats.to_csv("data/stats_clean.csv")
