==> 5_API.ipynb <==
import requests
import pandas as pd
import time

from datetime import date
from datetime import timedelta

pd.set_option("display.max_columns", None)

from functions import make_request
from functions import get_recent_games
from functions import clean_stats
from functions import aggregate_stats
from functions import get_stats

def get_team_code_map(df=False):
    # Make balldontlie api request and convert the json response to pandas dataframe
    team_code_df = make_request("teams")
    # team_code_df = team_code_df[["id", "city", "abbreviation", "full_name", "name"]]
    team_code_df = team_code_df.set_index("id")
    # using said dataframe, map team names to team id
    team_code_map = {}
    for row in team_code_df.iterrows():
        # Make sure "1" maps to 1. i.e. string maps to integer. 
        # This is so people can enter the team code in the 
        # text box for convenience and everything still works fine.
        team_code_map.update({str(row[0]): row[0]})
        # map the rest of the valid inputs to the team id 
        team_code_map.update(dict.fromkeys(row[1].str.lower().values, row[0]))                  
    if df:
        return team_code_df
    else:
        return team_code_map  # returning a dictionary

team_codes = get_team_code_map(df=True)
team_codes

team_df = make_request("teams")
team_df.to_dict(orient="records")

# Get a pandas dataframe of the 20 most recent games for the following team


recent_games = get_recent_games(1,2)
                                ### Should write a line of code here that drops the row if "time" is not an empty string (meaning the game is still in progress)

recent_games

game_ids_home, game_ids_away = get_recent_games("20","1")

def get_stats(game_ids_home, game_ids_away):
    """
    This function makes a request to balldontlie API for stats from specific games.
    The arguments for this function should be:
    1. a list of the 20 most recent game ids for the home team
    2. a list of the 20 most recent game ids for the away team
    
    The order matters. Putting the away team as the first argument and home team as the
    second will produce inaccurate results.
    
    The function returns a Numpy array that the model is expecting as input.
    """
    
    
    def format_params(game_ids):
        """ 
        Format query paramaters in a format the balldontlie API accepts
        e.g. ?game_ids[]=345686&game_ids[]=234356&gameids[]=3456356...
        """
        params = "game_ids[] " * len(game_ids)
        params = list(zip(params.split(" "), game_ids))
        params.append(("per_page", 100))
        return params
    
    stats_cols = ["ast","blk","dreb","fg3_pct","fg3a","fg3m","fg_pct","fga","fgm","ft_pct","fta","ftm","oreb",
              "pf","pts","reb","stl","turnover"]
    
    # Get pandas Series of home team stats
    params_home = format_params(game_ids_home)                                 # Get param list
    stats_home = make_request("stats", record_path="data", params=params_home) # Make request with said param list
    stats_home = clean_stats(stats_home)                                       # clean the data
    stats_home = stats_home[stats_home["team.id"].eq(stats_home["game.home_team_id"])]  # Filter for stats of players that played for the home team
    stats_home = aggregate_stats(stats_home)                                   # aggregate individual player stats into team stats
    stats_home = stats_home[stats_cols]                                        # Drop the columns that aren't basketball stats
    stats_home = stats_home.mean()                                             # average the stats
    
    # Get pandas Series of away team stats
    params_away = format_params(game_ids_away)
    stats_away = make_request("stats", record_path="data", params=params_away)
    stats_away = clean_stats(stats_away)
    stats_away = stats_away[stats_away["team.id"].eq(stats_away["game.visitor_team_id"])]
    stats_away = aggregate_stats(stats_away)
    stats_away = stats_away[stats_cols]
    stats_away = stats_away.mean()
    
    # Make a stats diff Series
    stats_diff = stats_home - stats_away
    
    # Rename columns and put it all together
    stats_home.index = "home_" + stats_home.index
    stats_away.index = "away_" + stats_away.index
    stats_diff.index = "diff_" + stats_diff.index
    
    stats = stats_home.append([stats_away, stats_diff])
    model_input = stats.values.reshape(1,-1)
    
    return model_input

stats = get_stats(game_ids_home, game_ids_away)

stats

### Testing the model

import pickle

model = pickle.load(open("model.sav", "rb"))

prediction = model.predict(stats)
prediction

# probability that away team will win
model.predict_proba(stats)[0][0]

# probability that home team will win
model.predict_proba(stats)[0][1]
